+++ src/config.cpp
@@ -731,6 +731,32 @@ BaseGameVersion ProjectConfig::stringToBaseGameVersion(const QString &string) {
     return version;
 }
 
+QString ProjectConfig::getPlayerIconPath(BaseGameVersion baseGameVersion, int character) {
+    switch (baseGameVersion) {
+        case BaseGameVersion::pokeemerald: {
+            static const QStringList paths = { QStringLiteral(":/icons/player/brendan_em.ico"),
+                                               QStringLiteral(":/icons/player/may_em.ico"), };
+            return paths.value(character);
+        }
+        case BaseGameVersion::pokefirered: {
+            static const QStringList paths = { QStringLiteral(":/icons/player/red.ico"),
+                                               QStringLiteral(":/icons/player/green.ico"), };
+            return paths.value(character);
+        }
+        case BaseGameVersion::pokeruby:    {
+            static const QStringList paths = { QStringLiteral(":/icons/player/brendan_rs.ico"),
+                                               QStringLiteral(":/icons/player/may_rs.ico"), };
+            return paths.value(character);
+        }
+        default: break;
+    }
+    return QString();
+}
+
+QIcon ProjectConfig::getPlayerIcon(BaseGameVersion baseGameVersion, int character) {
+    return QIcon(getPlayerIconPath(baseGameVersion, character));
+}
+
 ProjectConfig projectConfig;
 
 QString ProjectConfig::getConfigFilepath() {
@@ -868,6 +894,8 @@ void ProjectConfig::parseConfigKeyValue(QString key, QString value) {
         this->eventIconPaths[Event::Group::Heal] = value;
     } else if (key.startsWith("pokemon_icon_path/")) {
         this->pokemonIconPaths.insert(key.mid(QStringLiteral("pokemon_icon_path/").length()), value);
+    } else if (key == "events_tab_icon_path") {
+        this->eventsTabIconPath = value;
     } else if (key == "collision_sheet_path") {
         this->collisionSheetPath = value;
     } else if (key == "collision_sheet_width") {
@@ -890,6 +918,8 @@ void ProjectConfig::parseConfigKeyValue(QString key, QString value) {
             this->warpBehaviors.append(getConfigUint32(key, s));
     } else if (key == "max_events_per_group") {
         this->maxEventsPerGroup = getConfigInteger(key, value, 1, INT_MAX, 255);
+    } else if (key == "forced_major_version") {
+        this->forcedMajorVersion = getConfigInteger(key, value);
     } else {
         logWarn(QString("Invalid config key found in config file %1: '%2'").arg(this->getConfigFilepath()).arg(key));
     }
@@ -985,6 +1015,7 @@ QMap<QString, QString> ProjectConfig::getKeyValueMap() {
     for (auto it = this->identifiers.constBegin(); it != this->identifiers.constEnd(); it++) {
         map.insert("ident/"+defaultIdentifiers.value(it.key()).first, it.value());
     }
+    map.insert("events_tab_icon_path", this->eventsTabIconPath);
     map.insert("collision_sheet_path", this->collisionSheetPath);
     map.insert("collision_sheet_width", QString::number(this->collisionSheetSize.width()));
     map.insert("collision_sheet_height", QString::number(this->collisionSheetSize.height()));
@@ -997,6 +1028,7 @@ QMap<QString, QString> ProjectConfig::getKeyValueMap() {
         warpBehaviorStrs.append("0x" + QString("%1").arg(value, 2, 16, QChar('0')).toUpper());
     map.insert("warp_behaviors", warpBehaviorStrs.join(","));
     map.insert("max_events_per_group", QString::number(this->maxEventsPerGroup));
+    map.insert("forced_major_version", QString::number(this->forcedMajorVersion));
 
     return map;
 }