+++ src/mainwindow.cpp
@@ -250,7 +250,7 @@ void MainWindow::initCustomUI() {
 
     static const QMap<int, QIcon> mainTabIcons = {
         {MainTab::Map, QIcon(QStringLiteral(":/icons/minimap.ico"))},
-        {MainTab::Events, QIcon(QStringLiteral(":/icons/viewsprites.ico"))},
+        {MainTab::Events, ProjectConfig::getPlayerIcon(BaseGameVersion::pokefirered, 0)}, // Arbitrary default
         {MainTab::Header, QIcon(QStringLiteral(":/icons/application_form_edit.ico"))},
         {MainTab::Connections, QIcon(QStringLiteral(":/icons/connections.ico"))},
         {MainTab::WildPokemon, QIcon(QStringLiteral(":/icons/tall_grass.ico"))},
@@ -683,7 +683,7 @@ bool MainWindow::openProject(QString dir, bool initial) {
 
     // Make sure project looks reasonable before attempting to load it
     porysplash->showMessage("Verifying project");
-    if (!checkProjectSanity()) {
+    if (isInvalidProject(this->editor->project)) {
         delete this->editor->project;
         porysplash->stop();
         return false;
@@ -725,16 +725,20 @@ bool MainWindow::loadProjectData() {
     return success;
 }
 
-bool MainWindow::checkProjectSanity() {
-    if (editor->project->sanityCheck())
+bool MainWindow::isInvalidProject(Project *project) {
+    return !(checkProjectSanity(project) && checkProjectVersion(project));
+}
+
+bool MainWindow::checkProjectSanity(Project *project) {
+    if (project->sanityCheck())
         return true;
 
-    logWarn(QString("The directory '%1' failed the project sanity check.").arg(editor->project->root));
+    logWarn(QString("The directory '%1' failed the project sanity check.").arg(project->root));
 
     ErrorMessage msgBox(QStringLiteral("The selected directory appears to be invalid."), porysplash);
     msgBox.setInformativeText(QString("The directory '%1' is missing key files.\n\n"
                                       "Make sure you selected the correct project directory "
-                                      "(the one used to make your .gba file, e.g. 'pokeemerald').").arg(editor->project->root));
+                                      "(the one used to make your .gba file, e.g. 'pokeemerald').").arg(project->root));
     auto tryAnyway = msgBox.addButton("Try Anyway", QMessageBox::ActionRole);
     msgBox.exec();
     if (msgBox.clickedButton() == tryAnyway) {
@@ -743,6 +747,43 @@ bool MainWindow::checkProjectSanity() {
         return true;
     }
     return false;
+}
+
+bool MainWindow::checkProjectVersion(Project *project) {
+    QString error;
+    int projectVersion = project->getSupportedMajorVersion(&error);
+    if (projectVersion < 0) {
+        // Failed to identify a supported major version.
+        // We can't draw any conclusions from this, so we don't consider the project to be invalid.
+        QString msg = QStringLiteral("Failed to check project version");
+        logWarn(error.isEmpty() ? msg : QString("%1: '%2'").arg(msg).arg(error));
+    } else {
+        QString msg = QStringLiteral("Successfully checked project version. ");
+        logInfo(msg + ((projectVersion != 0) ? QString("Supports at least Porymap v%1").arg(projectVersion)
+                                             : QStringLiteral("Too old for any Porymap version")));
+
+        if (projectVersion < porymapVersion.majorVersion() && projectConfig.forcedMajorVersion < porymapVersion.majorVersion()) {
+            // We were unable to find the necessary changes for Porymap's current major version.
+            // Unless they have explicitly suppressed this message, warn the user that this might mean their project is missing breaking changes.
+            // Note: Do not report 'projectVersion' to the user in this message. We've already logged it for troubleshooting.
+            //       It is very plausible that the user may have reproduced the required changes in an
+            //       unknown commit, rather than merging the required changes directly from the base repo.
+            //       In this case the 'projectVersion' may actually be too old to use for their repo.
+            ErrorMessage msgBox(QStringLiteral("Your project may be incompatible!"), porysplash);
+            msgBox.setInformativeText(QString("Make sure '%1' has all the required changes for Porymap version %2."
+                                              "") // TODO: Once we have a wiki or manual page describing breaking changes, link that here.
+                                              .arg(project->getProjectTitle())
+                                              .arg(porymapVersion.majorVersion()));
+            auto tryAnyway = msgBox.addButton("Try Anyway", QMessageBox::ActionRole);
+            msgBox.exec();
+            if (msgBox.clickedButton() != tryAnyway){
+                return false;
+            }
+            // User opted to try with this version anyway. Don't warn them about this version again.
+            projectConfig.forcedMajorVersion = porymapVersion.majorVersion();
+        }
+    }
+    return true;
 }
 
 void MainWindow::showProjectOpenFailure() {
@@ -1148,11 +1189,6 @@ void MainWindow::displayMapProperties() {
     ui->frame_HeaderData->setEnabled(true);
     this->mapHeaderForm->setHeader(editor->map->header());
 
-    const QSignalBlocker b_PrimaryTileset(ui->comboBox_PrimaryTileset);
-    const QSignalBlocker b_SecondaryTileset(ui->comboBox_SecondaryTileset);
-    ui->comboBox_PrimaryTileset->setTextItem(editor->map->layout()->tileset_primary_label);
-    ui->comboBox_SecondaryTileset->setTextItem(editor->map->layout()->tileset_secondary_label);
-
     ui->mapCustomAttributesFrame->table()->setAttributes(editor->map->customAttributes());
 }
 
@@ -1265,6 +1301,21 @@ bool MainWindow::setProjectUI() {
     ui->layoutList->sortByColumn(0, Qt::SortOrder::AscendingOrder);
 
     ui->mapCustomAttributesFrame->table()->setRestrictedKeys(project->getTopLevelMapFields());
+
+    // Set a version dependent player icon (or user-chosen icon) for the Events tab.
+    QIcon eventTabIcon;
+    if (!projectConfig.eventsTabIconPath.isEmpty()) {
+        eventTabIcon = QIcon(project->getExistingFilepath(projectConfig.eventsTabIconPath));
+        if (eventTabIcon.isNull()) {
+            logWarn(QString("Failed to load custom Events tab icon '%1'.").arg(projectConfig.eventsTabIconPath));
+        }
+    }
+    if (eventTabIcon.isNull()) {
+        // We randomly choose between the available characters for ~flavor~.
+        // For now, this correctly assumes all versions have 2 icons.
+        eventTabIcon = ProjectConfig::getPlayerIcon(projectConfig.baseGameVersion, QRandomGenerator::global()->bounded(0, 2));
+    }
+    ui->mainTabBar->setTabIcon(MainTab::Events, eventTabIcon);
 
     return true;
 }